♻️ refactor: move favourite.py to extractor/utils/favourite.py
