+++ src/yutto/extractor/utils/favourite.py
